Update sl_ann.jl
Changed learning_rate to Float32(0.01)

+++ Series 05/Tutorial 05x08/sl_ann.jl
@@ -80,7 +80,7 @@ ps = params(model)
 
 # select optimizer
 
-learning_rate = 0.01
+learning_rate = Float32(0.01)
 
 opt = ADAM(learning_rate)
